clang-tidy error fixes

+++ tasks/konstantinov_s_elem_vec_sign_change_count/seq/src/ops_seq.cpp
@@ -1,11 +1,12 @@
 #include "konstantinov_s_elem_vec_sign_change_count/seq/include/ops_seq.hpp"
 
-#include <iostream>
-#include <numeric>
+//#include <iostream>
+//#include <numeric>
 #include <vector>
+#include <cstddef> 
 
 #include "konstantinov_s_elem_vec_sign_change_count/common/include/common.hpp"
-#include "util/include/util.hpp"
+//#include "util/include/util.hpp"
 
 namespace konstantinov_s_elem_vec_sign_change_count {
 
@@ -32,7 +33,7 @@ bool KonstantinovSElemVecSignChangeSEQ::RunImpl() {
   const auto invec = GetInput();
   int res = 0;
   for (size_t i = 0; i < invec.size() - 1; i++) {
-    res += (invec[i] > 0) != (invec[i + 1] > 0);  // + 1 если занки разные
+    res += static_cast<int>((invec[i] > 0) != (invec[i + 1] > 0));  // + 1 если занки разные
   }
   GetOutput() = res;
   return true;
